RAG/requirements.txt
fastapi 
uvicorn[standard] 
pymongo 
pinecone 
python-dotenv 
requests
langgraph
autogen
sentence-transformers
PyPDF2
python-multipart